issue: #355 issue: #142 issue: #113
Commenting code which is throwing message on exception on file change. This is causing crash of the tool, because now package is async and it can't be used with async package: here is exception from the tool on showing message:
System.InvalidOperationException: 'Due to high risk of deadlock you cannot call GetService from a background thread in an AsyncPackage derived class. You should instead call GetServiceAsync (without calling Result or Wait on the resultant Task object) or switch to the UI thread with the JoinableTaskFactory.SwitchToMainThreadAsync method before calling GetService.'

As well opening bundle configuration file is not useful at all, but only anoying.